Refactor device check
